quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16709 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzadmin/BZAdminClient.cxx
@@ -442,7 +442,7 @@ void BZAdminClient::outputServerList() const {
       if (!serverList.serverFound()) {
 	ui->outputMessage(std::string("...waiting on the list server..."), Yellow);
       } else {
-	ui->outputMessage(TextUtils::format("...retrieving list of servers... (found %d)", serverList.size()), Yellow);
+	ui->outputMessage(TextUtils::format("...retrieving list of servers... (found %d)", (int)serverList.size()), Yellow);
       }
     }
     serverList.checkEchos(&startupInfo);
